+++ mocks/pumps/MockPump.sol
@@ -1,6 +1,5 @@
 /**
  * SPDX-License-Identifier: MIT
- *
  */
 
 pragma solidity ^0.8.17;
@@ -14,15 +13,11 @@ import "src/interfaces/pumps/IPump.sol";
 contract MockPump is IPump {
     bytes public lastData;
 
-    function update(uint[] calldata, bytes calldata data) external {
+    function update(uint256[] calldata, bytes calldata data) external {
         lastData = data;
     }
 
-    function read(address, bytes calldata)
-        external
-        view
-        returns (bytes memory data)
-    {
+    function read(address, bytes calldata) external view returns (bytes memory data) {
         return lastData;
     }
 }
