+++ src/functions/ConstantProduct2.sol
@@ -18,9 +18,9 @@ import {LibMath} from "src/libraries/LibMath.sol";
  *  `b_i` is the reserve at index `i`
  */
 contract ConstantProduct2 is ProportionalLPToken2, IBeanstalkWellFunction {
-    using LibMath for uint;
+    using LibMath for uint256;
 
-    uint constant EXP_PRECISION = 1e12;
+    uint256 constant EXP_PRECISION = 1e12;
 
     /**
      * @dev `s = (b_0 * b_1)^(1/2)`
@@ -47,20 +47,20 @@ contract ConstantProduct2 is ProportionalLPToken2, IBeanstalkWellFunction {
      * >= 10^32.5, or about 100 trillion if tokens are measured to 18 decimal precision.
      */
     function calcLpTokenSupply(
-        uint[] calldata reserves,
+        uint256[] calldata reserves,
         bytes calldata
-    ) external pure override returns (uint lpTokenSupply) {
+    ) external pure override returns (uint256 lpTokenSupply) {
         lpTokenSupply = (reserves[0] * reserves[1] * EXP_PRECISION).sqrt();
     }
 
     /// @dev `b_j = s^2 / b_{i | i != j}`
     /// @dev rounds up
     function calcReserve(
-        uint[] calldata reserves,
-        uint j,
-        uint lpTokenSupply,
+        uint256[] calldata reserves,
+        uint256 j,
+        uint256 lpTokenSupply,
         bytes calldata
-    ) external pure override returns (uint reserve) {
+    ) external pure override returns (uint256 reserve) {
         // Note: potential optimization is to use unchecked math here
         reserve = lpTokenSupply ** 2;
         reserve = LibMath.roundUpDiv(reserve, reserves[j == 1 ? 0 : 1] * EXP_PRECISION);
@@ -77,12 +77,12 @@ contract ConstantProduct2 is ProportionalLPToken2, IBeanstalkWellFunction {
     /// @dev `b_j = (b_0 * b_1 * r_j / r_i)^(1/2)`
     /// Note: Always rounds down
     function calcReserveAtRatioSwap(
-        uint[] calldata reserves,
-        uint j,
-        uint[] calldata ratios,
+        uint256[] calldata reserves,
+        uint256 j,
+        uint256[] calldata ratios,
         bytes calldata
-    ) external pure override returns (uint reserve) {
-        uint i = j == 1 ? 0 : 1;
+    ) external pure override returns (uint256 reserve) {
+        uint256 i = j == 1 ? 0 : 1;
         // use 512 muldiv for last mul to avoid overflow
         reserve = (reserves[i] * reserves[j]).mulDiv(ratios[j], ratios[i]).sqrt();
     }
@@ -90,12 +90,12 @@ contract ConstantProduct2 is ProportionalLPToken2, IBeanstalkWellFunction {
     /// @dev `b_j = b_i * r_j / r_i`
     /// Note: Always rounds down
     function calcReserveAtRatioLiquidity(
-        uint[] calldata reserves,
-        uint j,
-        uint[] calldata ratios,
+        uint256[] calldata reserves,
+        uint256 j,
+        uint256[] calldata ratios,
         bytes calldata
-    ) external pure override returns (uint reserve) {
-        uint i = j == 1 ? 0 : 1;
+    ) external pure override returns (uint256 reserve) {
+        uint256 i = j == 1 ? 0 : 1;
         reserve = reserves[i] * ratios[j] / ratios[i];
     }
 }
